+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # CHANGELOG
 (Most recent on top.)
 
+### Upcoming:
+* Usable channel list.
+
+### 2017-11-02: v0.01.20 (@jws):
+* Added: Listing of posts containing a specified hashtag (last 20.)
+* Added: Create a message to send to a chatroom or private thread. Need to know the pre-existing channel number.
+* Added: Subscribed channel listing. Returns *only* channel number, owner/creator and the most recent message with its id. See also next item.
+* Added: Get a channel. Returns content for one post using code modified from above. Potentially useful as a subroutine?
+* Changed: Code: Reordered routines: post, reply and message pushed together.
+* Deprecated: -
+* Removed: pprint module import: I decided to examine JSON outside Pythonista.
+* Fixed: -
+* Security: -
+
 ### 2017-11-02: v0.01.19 (Threadsbared):
 * Added thread id to every displayed mention.
 * Reversed mentions sort order: most recent is last, at the bottom of the list, onscreen.
@@ -35,4 +49,3 @@
   * Repost a specific post number,
   * Get a post's content (raw response from server; not pretty).
 * Added documentation, but not yet /docs/.
-